Final Touches
Since all five steps were using the same CSS, I changed them to all have the same "step" class and moved all the SCSS code into "step.scss", where instead of copying the code 5 times for each step I just applied it all to the new step class.

I also added a link to the repo for this site to the footer. I think we mentioned doing that in class, but feel free to remove it if you don't think it belongs.